Split aio.rs to separate files.
This change doesn't modify any code, except for `use` and `mod`
statements.
