+++ redis/src/aio/mod.rs
@@ -1,0 +1,143 @@
+//! Adds async IO support to redis.
+use crate::cmd::{cmd, Cmd};
+use crate::connection::RedisConnectionInfo;
+use crate::types::{ErrorKind, RedisFuture, RedisResult, Value};
+#[cfg(all(not(feature = "tokio-comp"), feature = "async-std-comp"))]
+use ::async_std::net::ToSocketAddrs;
+use ::tokio::io::{AsyncRead, AsyncWrite};
+use async_trait::async_trait;
+use futures_util::Future;
+use std::net::SocketAddr;
+#[cfg(unix)]
+use std::path::Path;
+use std::pin::Pin;
+
+/// Enables the async_std compatibility
+#[cfg(feature = "async-std-comp")]
+#[cfg_attr(docsrs, doc(cfg(feature = "async-std-comp")))]
+pub mod async_std;
+
+/// Enables the tokio compatibility
+#[cfg(feature = "tokio-comp")]
+#[cfg_attr(docsrs, doc(cfg(feature = "tokio-comp")))]
+pub mod tokio;
+
+/// Represents the ability of connecting via TCP or via Unix socket
+#[async_trait]
+pub(crate) trait RedisRuntime: AsyncStream + Send + Sync + Sized + 'static {
+    /// Performs a TCP connection
+    async fn connect_tcp(socket_addr: SocketAddr) -> RedisResult<Self>;
+
+    // Performs a TCP TLS connection
+    #[cfg(any(feature = "tls-native-tls", feature = "tls-rustls"))]
+    async fn connect_tcp_tls(
+        hostname: &str,
+        socket_addr: SocketAddr,
+        insecure: bool,
+    ) -> RedisResult<Self>;
+
+    /// Performs a UNIX connection
+    #[cfg(unix)]
+    async fn connect_unix(path: &Path) -> RedisResult<Self>;
+
+    fn spawn(f: impl Future<Output = ()> + Send + 'static);
+
+    fn boxed(self) -> Pin<Box<dyn AsyncStream + Send + Sync>> {
+        Box::pin(self)
+    }
+}
+
+/// Trait for objects that implements `AsyncRead` and `AsyncWrite`
+pub trait AsyncStream: AsyncRead + AsyncWrite {}
+impl<S> AsyncStream for S where S: AsyncRead + AsyncWrite {}
+
+/// An async abstraction over connections.
+pub trait ConnectionLike {
+    /// Sends an already encoded (packed) command into the TCP socket and
+    /// reads the single response from it.
+    fn req_packed_command<'a>(&'a mut self, cmd: &'a Cmd) -> RedisFuture<'a, Value>;
+
+    /// Sends multiple already encoded (packed) command into the TCP socket
+    /// and reads `count` responses from it.  This is used to implement
+    /// pipelining.
+    fn req_packed_commands<'a>(
+        &'a mut self,
+        cmd: &'a crate::Pipeline,
+        offset: usize,
+        count: usize,
+    ) -> RedisFuture<'a, Vec<Value>>;
+
+    /// Returns the database this connection is bound to.  Note that this
+    /// information might be unreliable because it's initially cached and
+    /// also might be incorrect if the connection like object is not
+    /// actually connected.
+    fn get_db(&self) -> i64;
+}
+
+async fn authenticate<C>(connection_info: &RedisConnectionInfo, con: &mut C) -> RedisResult<()>
+where
+    C: ConnectionLike,
+{
+    if let Some(password) = &connection_info.password {
+        let mut command = cmd("AUTH");
+        if let Some(username) = &connection_info.username {
+            command.arg(username);
+        }
+        match command.arg(password).query_async(con).await {
+            Ok(Value::Okay) => (),
+            Err(e) => {
+                let err_msg = e.detail().ok_or((
+                    ErrorKind::AuthenticationFailed,
+                    "Password authentication failed",
+                ))?;
+
+                if !err_msg.contains("wrong number of arguments for 'auth' command") {
+                    fail!((
+                        ErrorKind::AuthenticationFailed,
+                        "Password authentication failed",
+                    ));
+                }
+
+                let mut command = cmd("AUTH");
+                match command.arg(password).query_async(con).await {
+                    Ok(Value::Okay) => (),
+                    _ => {
+                        fail!((
+                            ErrorKind::AuthenticationFailed,
+                            "Password authentication failed"
+                        ));
+                    }
+                }
+            }
+            _ => {
+                fail!((
+                    ErrorKind::AuthenticationFailed,
+                    "Password authentication failed"
+                ));
+            }
+        }
+    }
+
+    if connection_info.db != 0 {
+        match cmd("SELECT").arg(connection_info.db).query_async(con).await {
+            Ok(Value::Okay) => (),
+            _ => fail!((
+                ErrorKind::ResponseError,
+                "Redis server refused to switch database"
+            )),
+        }
+    }
+
+    Ok(())
+}
+
+mod connection;
+pub use connection::*;
+mod multiplexed_connection;
+pub use multiplexed_connection::*;
+#[cfg(feature = "connection-manager")]
+mod connection_manager;
+#[cfg(feature = "connection-manager")]
+pub use connection_manager::*;
+mod runtime;
+pub(super) use runtime::*;
